+++ src/data/mystery-encounters/encounters/clowning-around-encounter.ts
@@ -14,7 +14,7 @@ import { TrainerPartyTemplate } from "#app/data/trainers/TrainerPartyTemplate";
 import { ModifierTier } from "#app/modifier/modifier-tier";
 import type { PokemonHeldItemModifierType } from "#app/modifier/modifier-type";
 import { modifierTypes } from "#app/modifier/modifier-type";
-import { ModifierPoolType } from "#app/modifier/modifier-pool";
+import { ModifierPoolType } from "#app/modifier/modifier-pool-type";
 import { MysteryEncounterType } from "#enums/mystery-encounter-type";
 import { PartyMemberStrength } from "#enums/party-member-strength";
 import { globalScene } from "#app/global-scene";
