+++ src/modifier/modifier-pool-type.ts
@@ -1,0 +1,7 @@
+export enum ModifierPoolType {
+  PLAYER,
+  WILD,
+  TRAINER,
+  ENEMY_BUFF,
+  DAILY_STARTER,
+}
